+++ src/XrdCeph/XrdCephOss.cc
@@ -41,6 +41,8 @@
 #include "XrdCeph/XrdCephOssDir.hh"
 #include "XrdCeph/XrdCephOssFile.hh"
 
+#include <chrono>
+
 XrdVERSIONINFO(XrdOssGetStorageSystem, XrdCephOss);
 
 XrdSysError XrdCephEroute(0);
