+++ src/XrdCeph/XrdCephPosix.hh
@@ -37,6 +37,8 @@
 
 #include "XrdOuc/XrdOucIOVec.hh"
 
+#include "XrdCephBulkAioRead.hh"
+
 class XrdSfsAio;
 typedef void(AioCB)(XrdSfsAio*, size_t);
